Move old NavbarMain component to deprecated directory
The NavbarMainOld component has been renamed and moved to a new "~deprecated" directory to reflect that it's
