+++ frontend-next-migration/src/widgets/Navbar/index.ts
@@ -1,7 +1,7 @@
 // export { NavbarMainOld as Navbar } from './ui/NavbarMain/NavbarMainOld';
-export { NavbarClanMain as NavbarClan } from './ui/NavbarMain/NavbarMainOld';
-export { NavbarTeachingMain as NavbarTeaching } from './ui/NavbarMain/NavbarMainOld';
-export { NavbarGameArtMain as NavbarGameArt } from './ui/NavbarMain/NavbarMainOld';
+export { NavbarClanMain as NavbarClan } from './~deprecated/ui/NavbarMainOld';
+export { NavbarTeachingMain as NavbarTeaching } from './~deprecated/ui/NavbarMainOld';
+export { NavbarGameArtMain as NavbarGameArt } from './~deprecated/ui/NavbarMainOld';
 // export { NavbarMainAsync as Navbar} from "./ui/NavbarMain/NavbarMain.async";
